% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_red.R
\name{leer_red}
\alias{leer_red}
\title{Función de Lectrura de Red en ShapeFile}
\usage{
leer_red(ruta_red)
}
\arguments{
\item{ruta_red}{Es la ruta del archivo ShapeFile que corresponde a la red víal.}
}
\value{
retorna un objeto Simple Featrures tipo \code{LINESTRING}
}
\description{
Esta función lee un archivo ShapeFile con la estructura sf necesario para los siguientes análisis
}
