% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red2grafo.R
\name{red2grafo}
\alias{red2grafo}
\title{Crear un grafo con estructura Tidygraph}
\usage{
red2grafo(x, directed = TRUE)
}
\arguments{
\item{x}{Es la Red Vial en fomato \code{sf} (Simple Features), tipo \code{LINESTRING}}

\item{directed}{TRUE si la red es dirigida. Default es \code{TRUE}}
}
\value{
Grapfo cuya estructura contiene dos secciones una los \code{nodes} y
otra los \code{edges} pudiendo posteriormente acceder a elos de smanera
independiente para su manipulación y análisis
}
\description{
Esta funión crea un grafo con la estructura y formato de la librería
\href{https://cran.r-project.org/web/packages/tidygraph/index.html}{tidygraph}
desde un objeto Simple Features que correspnde a la Red Vial
}
